Add metrics from /metrics/snapshot endpoint for master and slaves
To gather more in-depth metrics, metrics from the default provided
`/metrics/snapshot` endpoint is now also collected. This is based
on the `/metrics/snapshot` response of Mesos 0.22.1

Master is just as important as the slaves, therefor now collecting
the master and slave metrics from the `/metrics/snapshot` endpoint.

The first scrape was only done after the first interval, this
is now fixed to also do an initial scrape on startup.

Change-Id: Ic45e2438946246396fe6888cd249932d432bc169